Add files via upload
This app is currently in progress, It includes anonymous live chat app that u can use. Definitely, I was a perfect match for people to seek connections with same situations and circumstances. You can write your thoughts on the diary and enjoy the pastel atmosphere of the app. Thanks for reading!